Switch JS and CSS bundling over to importmaps and dartcss
The dartcss part has been left out for now as the existing CSS is plain
CSS and doesn't need transpiling.

sorttable source code has been removed as it's kinda busted and doesn't
seem to be in use currently.

The bits of JS that used jQuery have been isolated in `bevan.js` and can
be converted over to ES6 modules as and when you'd like to do so.  Might
be worth turning them into Stimulus.js controllers.